# Conflicts:
#	nekoyume/Assets/Resources/Timeline/WorldBoss/Animations/AbyssFenrir/900001_Appear.playable
#	nekoyume/Assets/Resources/Timeline/WorldBoss/ContainerPrefabs/205007.prefab.meta
#	nekoyume/Assets/Resources/Worldboss/Prefab/205007_New.prefab.meta
#	nekoyume/Assets/Resources/Worldboss/Prefab/900001.prefab
#	nekoyume/Assets/Resources/Worldboss/Prefab/900001.prefab.meta
#	nekoyume/Assets/Resources/Worldboss/Prefab/900002.prefab
#	nekoyume/Assets/Resources/Worldboss/Prefab/900002.prefab.meta
#	nekoyume/Assets/_Scripts/Game/RaidStage.cs

+++ nekoyume/Assets/_Scripts/UI/Module/Timer/BattleTimerView.cs
@@ -19,7 +19,7 @@ namespace Nekoyume.UI.Module.Timer
             base.Show();
         }
 
-        public virtual void Show(double timeLimit)
+        public virtual void Show(int timeLimit)
         {
             _timeLimit = timeLimit;
             SetTimer(0);
